Fix plugin usage when {N} CLI is used as a library
The plugin hooks rely on `$options` object from CLI. This object is populated ONLY when CLI is used from command-line. However now we are able to build projects by requiring the CLI as a library.
In fact with `@next` version of CLI, the `release` and `bundle` options that the plugin uses from `$options` are available in the `hookArgs`. Use them and fallback to `$options` in case an older CLI (2.5.x for example) is used.